Convered every python script used in snakemake into a nps command
