+++ NPS_generation/__main__.py
@@ -2,7 +2,10 @@ import logging
 import argparse
 import os
 import NPS_generation
-from NPS_generation.commands import preprocess, create_training_sets, inner_train_models_RNN
+from NPS_generation.commands import (preprocess, create_training_sets, inner_train_models_RNN,
+                                     inner_sample_molecules_RNN,inner_tabulate_molecules,
+                                     inner_collect_tabulated_molecules, inner_process_tabulated_molecules,
+                                     inner_write_structural_prior_CV, inner_write_formula_prior_CV)
 
 
 logger = logging.getLogger("NPS_generation")
@@ -12,7 +15,9 @@ def main():
     parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument("--version", action="version", version=NPS_generation.__version__)
 
-    modules = (preprocess, create_training_sets, inner_train_models_RNN)
+    modules = (preprocess, create_training_sets, inner_train_models_RNN, inner_sample_molecules_RNN,
+               inner_tabulate_molecules, inner_collect_tabulated_molecules, inner_process_tabulated_molecules,
+               inner_write_structural_prior_CV, inner_write_formula_prior_CV)
 
     subparsers = parser.add_subparsers(title="Choose a command")
     subparsers.required = True
