+++ NPS_generation/commands/inner_sample_molecules_RNN.py
