+++ NPS_generation/commands/inner_tabulate_molecules.py
