+++ snakemake/Snakefile
@@ -171,7 +171,7 @@ rule sample_molecules_RNN:
         runtime=MODEL_PARAMS["sample_mols"]//10000,
         slurm_extra="--gres=gpu:1"
     shell:
-        'python ../NPS_generation/python/inner_sample_molecules_RNN.py '
+        'nps inner_sample_molecules_RNN '
         '--database {wildcards.dataset} '
         '--representation {wildcards.repr} '
         '--seed {wildcards.seed} '
@@ -202,7 +202,7 @@ rule tabulate_molecules:
         mem_mb=4000,
         runtime=MODEL_PARAMS["sample_mols"]//1000,
     shell:
-        'python ../NPS_generation/python/inner_tabulate_molecules.py '
+        'nps inner_tabulate_molecules '
         '--input_file {input.input_file} '
         '--train_file {input.train_file} '
         '--representation {wildcards.repr} '
@@ -222,7 +222,7 @@ rule collect_tabulated_molecules:
         mem_mb=4000,
         runtime=MODEL_PARAMS["sample_mols"]//1000,
     shell:
-        'python ../NPS_generation/python/inner_collect_tabulated_molecules.py '
+        'nps inner_collect_tabulated_molecules '
         '--input_files {input.input_files} '
         '--output_file {output.output_file} '
 
@@ -248,7 +248,7 @@ rule process_tabulated_molecules:
         mem_mb=4000,
         runtime=MODEL_PARAMS["sample_mols"]//1000,
     shell:
-        'python ../NPS_generation/python/inner_process_tabulated_molecules.py '
+        'nps inner_process_tabulated_molecules '
         '--input_file {input.input_file} '
         '--cv_file {input.cv_file} '
         '--output_file {output.output_file} '
@@ -285,7 +285,7 @@ rule write_structural_prior_CV:
         mem_mb=32000,
         runtime=180,
     shell:
-        'python ../NPS_generation/python/inner_write_structural_prior_CV.py '
+        'nps inner_write_structural_prior_CV '
         '--ranks_file {output.ranks_file} '
         '--tc_file {output.tc_file} '
         '--train_file {input.train_file} '
@@ -323,7 +323,7 @@ rule write_formula_prior_CV:
         mem_mb=32000,
         runtime=180,
     shell:
-        'python ../NPS_generation/python/inner_write_formula_prior_CV.py '
+        'nps inner_write_formula_prior_CV '
         '--ranks_file {output.ranks_file} '
         '--train_file {input.train_file} '
         '--test_file {input.test_file} '
@@ -363,7 +363,7 @@ rule write_structural_prior_CV_overall:
         mem_mb=32000,
         runtime=180,
     shell:
-        'python ../NPS_generation/python/inner_write_structural_prior_CV.py '
+        'nps inner_write_structural_prior_CV '
         '--ranks_file {output.ranks_file} '
         '--tc_file {output.tc_file} '
         '--train_file {input.train_file} '
